Run VMs with auto-create

+++ cloud_run.py
@@ -105,19 +105,26 @@ def gen_cloud_init(instance_id, local_hostname):
 
 def create_vm_img(os, size, name):
     vm_img_path = get_data_dir() / f"{name}.qcow2"
+    if vm_img_path.exists():
+        return vm_img_path, False
     shutil.copy(get_cloud_image(os), vm_img_path)
     subprocess.run(["qemu-img", "resize", vm_img_path, size], check=True)
-    return vm_img_path
+    return vm_img_path, True
 
 
-def create_vm(os, instance_id, local_hostname, mem, disk):
-    image = create_vm_img(os, disk, instance_id)
-    with gen_cloud_init(instance_id, local_hostname) as cloud_init:
-        subprocess.run([
-            "qemu-system-x86_64", 
-            "-m", mem,
-            "-nographic",
-            "-device", "virtio-net-pci,netdev=net0", "-netdev", "user,id=net0,hostfwd=tcp::2222-:22",
-            "-drive", f"if=virtio,format=qcow2,file={image}",
-            "-drive", f"if=virtio,format=raw,file={cloud_init}"
-        ], check=True)
+def run_vm(os, instance_id, local_hostname, mem, disk):
+    image, newly_created = create_vm_img(os, disk, instance_id)
+
+    qemu_command = [
+        "qemu-system-x86_64", 
+        "-m", mem,
+        "-nographic",
+        "-device", "virtio-net-pci,netdev=net0", "-netdev", "user,id=net0,hostfwd=tcp::2222-:22",
+        "-drive", f"if=virtio,format=qcow2,file={image}",
+    ]
+
+    if newly_created:
+        with gen_cloud_init(instance_id, local_hostname) as cloud_init:
+            subprocess.run(qemu_command + ["-drive", f"if=virtio,format=raw,file={cloud_init}"], check=True)
+    else:
+        subprocess.run(qemu_command, check=True)
